refactor(be): 중복 코드 제거 및 한 곳에서 관리 용이하게 구조 변경
