--- backend/src/main/java/com/deliveranything/domain/order/subscriber/delivery/DeliveryOfferedToRidersEventSubscriber.java
@@ -1,42 +1,0 @@
-package com.deliveranything.domain.order.subscriber.delivery;
-
-import com.deliveranything.domain.delivery.event.dto.DeliveryOfferedToRidersEvent;
-import com.deliveranything.domain.order.service.OrderService;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.annotation.PostConstruct;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.redis.connection.Message;
-import org.springframework.data.redis.connection.MessageListener;
-import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-import org.springframework.lang.NonNull;
-import org.springframework.stereotype.Component;
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class DeliveryOfferedToRidersEventSubscriber implements MessageListener {
-
-  private final RedisMessageListenerContainer container;
-  private final ObjectMapper objectMapper;
-  private final OrderService orderService;
-
-  @PostConstruct
-  public void registerListener() {
-    container.addMessageListener(this, new ChannelTopic("delivery-offered-to-riders-event"));
-  }
-
-  @Override
-  public void onMessage(@NonNull Message message, byte[] pattern) {
-    try {
-      log.info("라이더에게 상점이 수락한 주문이 뿌려짐");
-
-      DeliveryOfferedToRidersEvent event = objectMapper.readValue(message.getBody(),
-          DeliveryOfferedToRidersEvent.class);
-      orderService.processOrderTransmitted(event.orderId());
-    } catch (Exception e) {
-      log.error("Failed to process order assigned event from Redis", e);
-    }
-  }
-}
